Rename MemOnlyCache to ResizableCacheStore.
