+++ cs/samples/ResizableCacheStore/ISizeTracker.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-namespace MemOnlyCache
+namespace ResizableCacheStore
 {
     public interface ISizeTracker
     {
