+++ cs/samples/ResizableCacheStore/ResizableCacheStore.csproj
